refactor(map): Make custom cluster market component
